backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
PyJWT==2.10.1
email-validator==2.2.0
faiss-cpu==1.9.0.post1
numpy==2.1.3
pypdf==5.1.0
python-docx==1.1.2
python-pptx==1.0.2
openai==1.57.2
bcrypt==5.0.0
httpx==0.27.2
